:sparkles: Get basic Zora lister working

+++ src/Fleece.sol
@@ -14,8 +14,9 @@ library Fleece {
     using JsonWriter for JsonWriter.Json;
 
     uint8 internal constant MAX_NUMBER_OF_ESSAY_ELEMENTS = 23;
-    
+
     uint8 internal constant RETURN_SUCCESS = 0;
+
     // uint8 internal constant RETURN_ERROR_INVALID_JSON = 1;
     // uint8 internal constant RETURN_ERROR_PART = 2;
     // uint8 internal constant RETURN_ERROR_NO_MEM = 3;
@@ -30,22 +31,27 @@ library Fleece {
         require(returnValue == RETURN_SUCCESS, "JSON parsing error, you have been slayed by the hydra");
 
         // TODO investigate type casting messiness here
-        return Essay(
-            uint8(uint(JsmnSolLib.parseInt(readJsonElement(_json, tokens, 2)))),
-            uint8(uint(JsmnSolLib.parseInt(readJsonElement(_json, tokens, 4)))),
-            uint16(uint(JsmnSolLib.parseInt(readJsonElement(_json, tokens, 6)))),
-            readJsonElement(_json, tokens, 8),
-            readJsonElement(_json, tokens, 10),
-            readJsonElement(_json, tokens, 12),
-            readJsonElement(_json, tokens, 14),
-            readJsonElement(_json, tokens, 16),
-            readJsonElement(_json, tokens, 18),
-            readJsonElement(_json, tokens, 20),
-            readJsonElement(_json, tokens, 22)
-        );
+        return
+            Essay(
+                uint8(uint256(JsmnSolLib.parseInt(readJsonElement(_json, tokens, 2)))),
+                uint8(uint256(JsmnSolLib.parseInt(readJsonElement(_json, tokens, 4)))),
+                uint16(uint256(JsmnSolLib.parseInt(readJsonElement(_json, tokens, 6)))),
+                readJsonElement(_json, tokens, 8),
+                readJsonElement(_json, tokens, 10),
+                readJsonElement(_json, tokens, 12),
+                readJsonElement(_json, tokens, 14),
+                readJsonElement(_json, tokens, 16),
+                readJsonElement(_json, tokens, 18),
+                readJsonElement(_json, tokens, 20),
+                readJsonElement(_json, tokens, 22)
+            );
     }
 
-    function readJsonElement(string memory _json, JsmnSolLib.Token[] memory _elements, uint256 _index) private pure returns (string memory) {
+    function readJsonElement(
+        string memory _json,
+        JsmnSolLib.Token[] memory _elements,
+        uint256 _index
+    ) private pure returns (string memory) {
         return JsmnSolLib.getBytes(_json, _elements[_index].start, _elements[_index].end);
     }
 
@@ -65,7 +71,7 @@ library Fleece {
         writer = writer.writeStringProperty("Publication URL", _essay.publicationURL);
         writer = writer.writeStringProperty("Archival URL", _essay.archivalURL);
         writer = writer.writeEndObject();
-        
+
         return writer.value;
-    }    
+    }
 }
